requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
aiosignal==1.3.1
moralis==0.1.43
psycopg2-binary
